Fix all 26 build errors in plumbing subsystem (Phase 179)
- PlumbingWaterSafetyCommands: qualify ParseDate/ParseDateStr with
  WaterSafetyDateHelper class prefix; add using Planscape.Docs.Templates
  for MiniWordAdapter
- RealTimePipeSizer: split Register into UIControlledApplication /
  UIApplication overloads sharing a private RegisterCore() so
  StingCommandHandler can call it from Execute(UIApplication)
- PlumbingVisualisationCommands: remove unused zoneLabel variable and
  all its assignments (CS0219)
- PlumbingSpoolCommands: fix AssemblyInstance.IsValidNamingCategory to
  pass ids as third argument; replace removed Revit 2025 API
  (ScheduleField.ParameterName, SchedulableField.GetSchedulableFieldName)
  with ParameterId-based BIP field deduplication
- PipeNetworkGraph: replace non-existent MEPModel.GetMEPSystems() with
  connector-based system name lookup
- DrainageSchematicGenerator: add explicit (int) casts for node.DnMm/2
  double-to-int assignments
- TMVEngine: PLM_TMV_ANNUAL_DUE → PLM_TMV_NEXT_TEST (already applied)
- PlumbingNetworkCommands: VentPipesCreated/Skipped → VentsCreated/
  VentsSkipped (already applied)

+++ StingTools/Core/Plumbing/DrainageSchematicGenerator.cs
@@ -430,7 +430,7 @@ namespace StingTools.Core.Plumbing
                 int floorCount = Math.Max(2, (int)Math.Ceiling(heightFt / (opts.LevelHeightMm * MmToFt)));
 
                 // Vent sizing (roughly half stack DN)
-                int ventDnMm = node.DnMm > 0 ? Math.Max(50, node.DnMm / 2) : 50;
+                int ventDnMm = node.DnMm > 0 ? Math.Max(50, (int)(node.DnMm / 2)) : 50;
 
                 var sl = new StackLayout
                 {
@@ -487,7 +487,7 @@ namespace StingTools.Core.Plumbing
                         {
                             FloorIndex   = f,
                             FixtureCount = 0,
-                            BranchDnMm   = Math.Max(50, node.DnMm / 2),
+                            BranchDnMm   = Math.Max(50, (int)(node.DnMm / 2)),
                             SlopePct     = 1.25,
                             FixtureLabel = "",
                             IsLeft       = f % 2 != 0
